pyproject.toml
[tool.poetry]
name = "drft"
version = "0.0.0-alpha.1"
description = ""
authors = ["Anthony Almarza <anthony.almarza@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "drft", from = "src" },
]
# https://pypi.org/classifiers/
classifiers = [
    "Development Status :: 3 - Alpha",
    #"Development Status :: 4 - Beta",
    #"Development Status :: 5 - Production/Stable",
    #"Development Status :: 6 - Mature",
    "Natural Language :: English",
    "Programming Language :: Python :: 3 :: Only",
    "Topic :: Software Development :: Libraries :: Python Modules"
]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.0"
djangorestframework = "^3.11"
django-filter = "^2.2.0"
drf-yasg = { version = "^1.17", optional = true }

[tool.poetry.extras]
yasg = ["drf-yasg"]

[tool.poetry.dev-dependencies]
ipython = "^7.19.0"
pre-commit = "^2.9.3"
black = "^20.8b1"
flake8 = "^3.8.4"
mypy = "^0.812"
pylint = "^2.7.2"
pylint-django = "^2.4.2"
pyproject-script = "^0.0.0-alpha.1"
pytest = "^6"
pytest-sugar = "^0.9.4"
pytest-mock = "^3.4.0"
pytest-responses = "^0.4.0"
pytest-picked = "^0.4.4"
pytest-instafail = "^0.4.2"
pytest-django = "^4.1.0"
pytest-cov = "^2.11.1"
freezegun = "^1.0.0"
factory-boy = "^3.1.0"
codecov = "^2.1.11"
coverage = {extras = ["toml"], version = "^5.3"}
django-coverage-plugin = "^1.8.0"
tox = "^3.23.0"
bandit = "^1.7.0"
Sphinx = "^3.5.3"
sphinx-sitemap = "^2.2.0"
sphinx-pyreverse = "^0.0.14"
sphinx-autodoc-typehints = "^1.11.1"
sphinx-autobuild = "^2021.3.14"
recommonmark = "^0.7.1"
isort = "^5.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.run]
source = [
    "drft",
]
omit = [
    "tests/*",
    "factories/*"
]

[tool.coverage.report]
omit = [
    "tests/*",
    "factories/*"
]
skip_empty = true
show_missing = true
#fail_under = 70

[tool.black]
line-length = 79

[tool.isort]
profile = "black"
line_length = 79

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tests.settings"
addopts = "-s --cov=src/drft --no-cov-on-fail"
python_files = [
    "test_*.py",
    "tests.py",
    "*_tests.py"
]
testpaths = [
    "./tests",
]
